move predictors around
