+++ cloud/src/autogluon/cloud/predictor/__init__.py
@@ -1,6 +1,4 @@
-from .predictor import (
-    TabularCloudPredictor,
-    TextCloudPredictor,
-    ImageCloudPredictor,
-    MultiModalCloudPredictor,
-)
+from .tabular_cloud_predictor import TabularCloudPredictor
+from .text_cloud_predictor import TextCloudPredictor
+from .image_cloud_predictor import ImageCloudPredictor
+from .multimodal_cloud_predictor import MultiModalCloudPredictor
