+++ cloud/src/autogluon/cloud/predictor/cloud_predictor.py
@@ -979,332 +979,3 @@ class CloudPredictor(ABC):
             predictor._endpoint_saved = None
         # TODO: Version compatibility check
         return predictor
-
-
-class TabularCloudPredictor(CloudPredictor):
-
-    predictor_file_name = 'TabularCloudPredictor.pkl'
-
-    @property
-    def predictor_type(self):
-        return 'tabular'
-
-    def _get_local_predictor_cls(self):
-        from autogluon.tabular import TabularPredictor
-        predictor_cls = TabularPredictor
-        return predictor_cls
-
-    def _construct_config(self, predictor_init_args, predictor_fit_args, leaderboard, **kwargs):
-        assert self.predictor_type is not None
-        if 'feature_metadata' in predictor_fit_args:
-            predictor_fit_args = copy.deepcopy(predictor_fit_args)
-            feature_metadata = predictor_fit_args.pop('feature_metadata')
-            feature_metadata = dict(
-                type_map_raw=feature_metadata.type_map_raw,
-                type_map_special=feature_metadata.get_type_map_special(),
-            )
-            assert 'feature_metadata' not in kwargs, 'feature_metadata in both `predictor_fit_args` and kwargs. This should not happen.'
-            kwargs['feature_metadata'] = feature_metadata
-        config = dict(
-            predictor_type=self.predictor_type,
-            predictor_init_args=predictor_init_args,
-            predictor_fit_args=predictor_fit_args,
-            leaderboard=leaderboard,
-            **kwargs
-        )
-        path = os.path.join(self.local_output_path, 'utils', 'config.yaml')
-        with open(path, 'w') as f:
-            yaml.dump(config, f)
-        return path
-
-    def predict_real_time(self, test_data, test_data_image_column=None, accept='application/x-parquet'):
-        """
-        Predict with the deployed SageMaker endpoint. A deployed SageMaker endpoint is required.
-        This is intended to provide a low latency inference.
-        If you want to inference on a large dataset, use `predict()` instead.
-
-        Parameters
-        ----------
-        test_data: Union(str, pandas.DataFrame)
-            The test data to be inferenced.
-            Can be a pandas.DataFrame, a local path or a s3 path to a csv file.
-            When predicting multimodality with image modality:
-                You need to specify `test_data_image_column`, and make sure the image column contains relative path to the image.
-        test_data_image_column: Optional(str)
-            If provided a pandas.DataFrame as the test_data and test_data involves image modality,
-            you must specify the column name corresponding to image paths.
-            Images have to live in the same directory specified by the column.
-        accept: str, default = application/x-parquet
-            Type of accept output content.
-            Valid options are application/x-parquet, text/csv, application/json
-
-        Returns
-        -------
-        Pandas.DataFrame
-        Predict results in DataFrame
-        """
-        assert self.endpoint, 'Please call `deploy()` to deploy an endpoint first.'
-        assert accept in VALID_ACCEPT, f'Invalid accept type. Options are {VALID_ACCEPT}.'
-
-        if isinstance(test_data, str):
-            test_data = load_pd.load(test_data)
-        if isinstance(test_data, pd.DataFrame):
-            if test_data_image_column is not None:
-                test_data = convert_image_path_to_encoded_bytes_in_dataframe(test_data, test_data_image_column)
-
-        return self._predict_real_time(test_data=test_data, accept=accept)
-
-
-class TextCloudPredictor(CloudPredictor):
-
-    predictor_file_name = 'TextCloudPredictor.pkl'
-
-    @property
-    def predictor_type(self):
-        return 'text'
-
-    def _get_local_predictor_cls(self):
-        from autogluon.text import TextPredictor
-        predictor_cls = TextPredictor
-        return predictor_cls
-
-
-class ImageCloudPredictor(CloudPredictor):
-
-    predictor_file_name = 'ImageCloudPredictor.pkl'
-
-    @property
-    def predictor_type(self):
-        return 'image'
-
-    @property
-    def _realtime_predictor_cls(self):
-        return AutoGluonImageRealtimePredictor
-
-    def _get_local_predictor_cls(self):
-        from autogluon.vision import ImagePredictor
-        predictor_cls = ImagePredictor
-        return predictor_cls
-
-    def fit(
-        self,
-        *,
-        predictor_init_args,
-        predictor_fit_args,
-        image_path,
-        **kwargs
-    ):
-        super().fit(
-            predictor_init_args=predictor_init_args,
-            predictor_fit_args=predictor_fit_args,
-            image_path=image_path,
-            **kwargs
-        )
-
-    def predict_real_time(self, test_data, test_data_image_column=None, accept='application/x-parquet'):
-        """
-        Predict with the deployed SageMaker endpoint. A deployed SageMaker endpoint is required.
-        This is intended to provide a low latency inference.
-        If you want to inference on a large dataset, use `predict()` instead.
-
-        Parameters
-        ----------
-        test_data: Union(str, pandas.DataFrame)
-            The test data to be inferenced.
-            Can be a pandas.DataFrame, a numpy.ndarray, a local path or a s3 path to a csv file containing paths of test images.
-            Or a local path to a single image file.
-            Or a list of local paths to image files.
-        test_data_image_column: Optional(str)
-            If provided a pandas.DataFrame as the test_data, you must specify the column name corresponding to image paths.
-            Images have to live in the same directory specified by the column.
-        accept: str, default = application/x-parquet
-            Type of accept output content.
-            Valid options are application/x-parquet, text/csv, application/json
-
-        Returns
-        -------
-        Pandas.DataFrame
-        Predict results in DataFrame
-        """
-        assert self.endpoint, 'Please call `deploy()` to deploy an endpoint first.'
-        assert accept in VALID_ACCEPT, f'Invalid accept type. Options are {VALID_ACCEPT}.'
-
-        import numpy as np
-
-        if isinstance(test_data, str):
-            image = cv2.imread(test_data)
-            if image is None:  # not an image
-                test_data = load_pd.load(test_data)
-            else:
-                test_data = image
-        if isinstance(test_data, list):
-            images = []
-            for image in test_data:
-                images.append(cv2.imread(image))
-            test_data = np.array(images, dtype=object)
-        if isinstance(test_data, pd.DataFrame):
-            assert test_data_image_column is not None, 'Please specify an image column name'
-            assert test_data_image_column in test_data, 'Please specify a valid image column name'
-
-            # Convert test data to be numpy array for network transfer
-            test_data = np.asarray([cv2.imread(path) for path in test_data[test_data_image_column]])
-        assert isinstance(test_data, np.ndarray), f'Invalid test data format {type(test_data)}'
-
-        return self._predict_real_time(test_data=test_data, accept=accept)
-
-    def predict(
-        self,
-        test_data,
-        **kwargs,
-    ):
-        """
-        test_data: str
-            The test data to be inferenced. Can be a local path or a s3 path to a directory containing the images.
-        kwargs:
-            Refer to `CloudPredictor.predict()`
-        """
-        split_type = None
-        content_type = 'application/x-image'
-        kwargs = copy.deepcopy(kwargs)
-        transformer_kwargs = kwargs.pop('transformer_kwargs', dict())
-        transformer_kwargs['strategy'] = 'SingleRecord'
-        super().predict(
-            test_data,
-            split_type=split_type,
-            content_type=content_type,
-            transformer_kwargs=transformer_kwargs,
-            **kwargs,
-        )
-
-
-class MultiModalCloudPredictor(CloudPredictor):
-
-    predictor_file_name = 'MultiModalCloudPredictor.pkl'
-
-    @property
-    def predictor_type(self):
-        return 'multimodal'
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    @property
-    def _realtime_predictor_cls(self):
-        return AutoGluonMultiModalRealtimePredictor
-
-    def _get_local_predictor_cls(self):
-        import autogluon.text
-        from distutils.version import LooseVersion
-
-        if LooseVersion(autogluon.text.__version__) < LooseVersion('0.5'):
-            from autogluon.text.automm import AutoMMPredictor
-            multimodal_predictor_cls = AutoMMPredictor
-        else:
-            from autogluon.multimodal import MultiModalPredictor
-            multimodal_predictor_cls = MultiModalPredictor
-
-        predictor_cls = multimodal_predictor_cls
-        return predictor_cls
-
-    def predict_real_time(self, test_data, test_data_image_column=None, accept='application/x-parquet'):
-        """
-        Predict with the deployed SageMaker endpoint. A deployed SageMaker endpoint is required.
-        This is intended to provide a low latency inference.
-        If you want to inference on a large dataset, use `predict()` instead.
-
-        Parameters
-        ----------
-        test_data: Union(str, pandas.DataFrame)
-            The test data to be inferenced.
-            Can be a pandas.DataFrame, a local path or a s3 path to a csv file.
-            When predicting multimodality with image modality:
-                You need to specify `test_data_image_column`, and make sure the image column contains relative path to the image.
-            When predicting with only images:
-                Can be a pandas.DataFrame, a local path or a s3 path to a csv file.
-                    Similarly, You need to specify `test_data_image_column`, and make sure the image column contains relative path to the image.
-                Or a local path to a single image file.
-                Or a list of local paths to image files.
-        test_data_image_column: Optional(str)
-            If provided a pandas.DataFrame as the test_data and test_data involves image modality,
-            you must specify the column name corresponding to image paths.
-            Images have to live in the same directory specified by the column.
-        accept: str, default = application/x-parquet
-            Type of accept output content.
-            Valid options are application/x-parquet, text/csv, application/json
-
-        Returns
-        -------
-        Pandas.DataFrame
-        Predict results in DataFrame
-        """
-        assert self.endpoint, 'Please call `deploy()` to deploy an endpoint first.'
-        assert accept in VALID_ACCEPT, f'Invalid accept type. Options are {VALID_ACCEPT}.'
-
-        import numpy as np
-
-        if isinstance(test_data, str):
-            image = cv2.imread(test_data)
-            if image is None:  # not an image
-                test_data = load_pd.load(test_data)
-            else:
-                test_data = np.array([read_image_bytes_and_encode(test_data)], dtype='object')
-                content_type = 'application/x-npy'
-        if isinstance(test_data, list):
-            images = []
-            test_data = np.array([read_image_bytes_and_encode(image) for image in images], dtype='object')
-            content_type = 'application/x-npy'
-        if isinstance(test_data, pd.DataFrame):
-            if test_data_image_column is not None:
-                test_data = convert_image_path_to_encoded_bytes_in_dataframe(test_data, test_data_image_column)
-            content_type = 'application/x-parquet'
-
-        # Providing content type here because sagemaker serializer doesn't support change content type dynamically.
-        # Pass to `endpoint.predict()` call as `initial_args` instead
-        return self._predict_real_time(test_data=test_data, accept=accept, ContentType=content_type)
-
-    def predict(
-        self,
-        test_data,
-        test_data_image_column=None,
-        image_modality_only=False,
-        **kwargs,
-    ):
-        """
-        test_data: str
-            The test data to be inferenced.
-            Can be a pandas.DataFrame, a local path or a s3 path to a csv file.
-            When predicting multimodality with image modality:
-                You need to specify `test_data_image_column`, and make sure the image column contains relative path to the image.
-            When predicting with only images:
-                User has to specify `image_modality_only` for this mode.
-                Can be a local path or a s3 path to a directory containing the images.
-        test_data_image_column: Optional(str)
-            If test_data involves image modality, you must specify the column name corresponding to image paths.
-            Images have to live in the same directory specified by the column.
-        image_modality_only: Bool
-            Whether the test_data only contains image modality.
-            If so, test data can be a local path or a s3 path to a directory containing the images.
-            Otherwise, you can only provide a pandas.DataFrame, a local path or a s3 path to a csv file, and specify `test_data_image_column`.
-        kwargs:
-            Refer to `CloudPredictor.predict()`
-        """
-        if image_modality_only:
-            split_type = None
-            content_type = 'application/x-image'
-            kwargs = copy.deepcopy(kwargs)
-            transformer_kwargs = kwargs.pop('transformer_kwargs', dict())
-            transformer_kwargs['strategy'] = 'SingleRecord'
-            super().predict(
-                test_data,
-                test_data_image_column=test_data_image_column,
-                split_type=split_type,
-                content_type=content_type,
-                transformer_kwargs=transformer_kwargs,
-                **kwargs,
-            )
-        else:
-            super().predict(
-                test_data,
-                test_data_image_column=test_data_image_column,
-                **kwargs,
-            )
